requirements.txt
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
Pillow==10.0.1
PyAudio==0.2.13
PyAutoGUI==0.9.53
python_dateutil==2.8.2
tornado==6.3.2
